apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eidolon-api"
version = "0.1.0"
description = "BURCH-EIDOLON FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "pydantic-settings>=2.5.2",
  "httpx>=0.27.0",
  "numpy>=2.1.1",
  "python-dateutil>=2.9.0",
  "reportlab>=4.2.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0"
]

[tool.setuptools.packages.find]
include = ["eidolon_api*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
